+++ src/Generating/ProtIntGen.h
@@ -767,7 +767,8 @@ public:
 					if (IsBiomeOcean(above) || IsBiomeOcean(below) || IsBiomeOcean(left) || IsBiomeOcean(right))
 					{
 						// First convert the value to a regular biome (drop the M flag), then modulo by our biome count:
-						val = ToBeach[(val % 128) % ARRAYCOUNT(ToBeach)];
+						// KINGCOL13TODO: work out same as above in IntGen
+						val = ToBeach[static_cast<unsigned long>(val % 128) % ARRAYCOUNT(ToBeach)];
 					}
 				}
 				a_Values[x + z * a_SizeX] = val;
